Infer values with a length less than 4 as a string
python-dateutil parses things like "T4" or "T10" as a datetime. These can be identifiers referencing areas of the spine, for one example. Additionally, it is very unlikely that we'll see values with a length less than 4 that are representative as a date value, and even then falling back to string should be ok.